Fixes base path issue
Fixes issue when you try to save to /foo/bar and the URL is
/media/cache/baz.jpg it will actually save into
/foo/bar/media/cache/baz.jpg when the expected result was
/foo/bar/cache/baz/jpg since the cache_base_path configuration was set.

This _may_ of been caused by commit
07228a5d7bf1382a1843c3807e371853bf98c754 which fixed the app_*.php
base URL issues. The baseUrl was set to the basePath variable which was
overriding the cache_base_path configuration in all situations which
made the configuration pointless.

The solution was to set the request's baseUrl into a new variable
baseUrl instead of the basePath therefore the basePath will be able to
be configured.